# openapi.yaml
openapi: 3.0.3
info:
  title: Sistem Pendataan Inventaris Laboratorium REST API
  description: |
    API Spesifikasi untuk Sistem Pendataan Inventaris Laboratorium (Sistem Pendataan Inventaris Laboratorium) 
    menggunakan Spring Boot, JWT Authentication, dan Role-Based Access Control (RBAC).
    
    ### Peran Pengguna (Roles):
    - **ADMIN**: Memiliki akses penuh ke seluruh sistem termasuk CRUD kategori, barang, laporan, dan aksi peminjaman.
    - **PETUGAS**: Memiliki hak akses serupa dengan ADMIN untuk operasional sehari-hari (CRUD kategori/barang, persetujuan peminjaman, laporan).
    - **MAHASISWA**: Hanya dapat melihat kategori/barang, melakukan pengajuan peminjaman barang untuk diri sendiri, dan melihat riwayat peminjaman miliknya sendiri.
    
    ### Alur Autentikasi:
    1. Lakukan registrasi melalui `/api/auth/register` (opsional jika menggunakan akun seeder bawaan).
    2. Lakukan login melalui `/api/auth/login` untuk mendapatkan token JWT.
    3. Masukkan token tersebut dalam Header HTTP: `Authorization: Bearer <JWT_TOKEN>` untuk mengakses endpoint yang dilindungi.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Registrasi Pengguna Baru
      description: Mendaftarkan akun pengguna baru dengan menetapkan nama, email, password, dan perannya (ADMIN, PETUGAS, MAHASISWA).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registrasi berhasil dilakukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseUserResponse'
        '400':
          description: Data input tidak valid atau email sudah terdaftar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseValidationError'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login Pengguna
      description: Autentikasi menggunakan email dan password untuk mendapatkan JWT Token untuk akses API selanjutnya.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login berhasil, JWT token dikembalikan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseJwtResponse'
        '401':
          description: Email atau password salah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '400':
          description: Format data input tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseValidationError'
...
  /api/barang:
    get:
      tags:
        - Barang Inventaris
      summary: Dapatkan Semua Barang Inventaris
      description: Mengambil daftar barang dengan sistem paginasi, pengurutan, dan info relasi kategori. Dapat diakses oleh ADMIN, PETUGAS, dan MAHASISWA.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Nomor halaman (0-indexed)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Jumlah data per halaman
          required: false
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: Kolom/properti untuk pengurutan
          required: false
          schema:
            type: string
            default: idBarang
        - name: sortDir
          in: query
          description: Arah pengurutan (ASC atau DESC)
          required: false
          schema:
            type: string
            enum: [ASC, DESC]
            default: ASC
      responses:
        '200':
          description: Berhasil memuat daftar barang
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseListBarangResponse'
        '401':
          description: Token tidak valid atau tidak disertakan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'

    post:
      tags:
        - Barang Inventaris
      summary: Daftarkan Barang Baru
      description: Menambahkan barang inventaris baru ke sistem. Menghubungkan ke kategori tertentu. Hanya dapat diakses oleh **ADMIN** dan **PETUGAS**.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BarangRequest'
      responses:
        '201':
          description: Barang berhasil didaftarkan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseBarangResponse'
        '400':
          description: Validasi gagal / Kode barang sudah terdaftar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseValidationError'
        '404':
          description: ID Kategori tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '403':
          description: Hak akses ditolak (Hanya ADMIN/PETUGAS)
  /api/barang/{id}:
    get:
      tags:
        - Barang Inventaris
      summary: Detail Barang berdasarkan ID
      description: Mengambil informasi detail satu barang inventaris. Dapat diakses oleh ADMIN, PETUGAS, dan MAHASISWA.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID Barang
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Barang ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseBarangResponse'
        '404':
          description: Barang tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'

    put:
      tags:
        - Barang Inventaris
      summary: Perbarui Barang
      description: Memperbarui data barang inventaris berdasarkan ID. Hanya dapat diakses oleh **ADMIN** dan **PETUGAS**.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID Barang yang akan diperbarui
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BarangRequest'
      responses:
        '200':
          description: Barang berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseBarangResponse'
        '404':
          description: Barang atau Kategori tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '400':
          description: Validasi gagal / Kode barang duplikat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseValidationError'
        '403':
          description: Hak akses ditolak

    delete:
      tags:
        - Barang Inventaris
      summary: Hapus Barang
      description: Menghapus data barang inventaris berdasarkan ID. Hanya dapat diakses oleh **ADMIN** dan **PETUGAS**.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID Barang yang akan dihapus
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Barang berhasil dihapus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '404':
          description: Barang tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '403':
          description: Hak akses ditolak
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Masukkan token JWT Anda dalam format `Bearer <token>` pada header Otorisasi HTTP.
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - nama
        - email
        - password
        - role
      properties:
        nama:
          type: string
          minLength: 2
          maxLength: 100
          example: Ahmad Mahasiswa
        email:
          type: string
          format: email
          example: mahasiswa@lab.com
        password:
          type: string
          minLength: 6
          example: mahasiswa123
        role:
          $ref: '#/components/schemas/Role'
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@lab.com
        password:
          type: string
          example: admin123
...
    BarangRequest:
      type: object
      required:
        - idKategori
        - namaBarang
        - kodeBarang
        - jumlahTotal
        - kondisi
        - lokasi
      properties:
        idKategori:
          type: integer
          format: int64
          example: 1
        namaBarang:
          type: string
          example: Digital Multimeter Sanwa CD800a
        kodeBarang:
          type: string
          example: BRG-EL-001
        jumlahTotal:
          type: integer
          minimum: 1
          example: 15
        kondisi:
          $ref: '#/components/schemas/KondisiBarang'
        lokasi:
          type: string
          example: Lab Elektronika, Lemari B, Rak 3
...
    Role:
      type: string
      enum:
        - ADMIN
        - PETUGAS
        - MAHASISWA
      example: ADMIN
    KondisiBarang:
      type: string
      enum:
        - BAIK
        - RUSAK_RINGAN
        - RUSAK_BERAT
      example: BAIK
...
    WebResponseString:
      type: object
      properties:
        code:
          type: integer
          example: 200
        status:
          type: string
          example: OK
        message:
          type: string
          example: Operasi berhasil!
        data:
          type: string
          nullable: true
          example: null
    WebResponseValidationError:
      type: object
      properties:
        code:
          type: integer
          example: 400
        status:
          type: string
          example: BAD REQUEST
        message:
          type: string
          example: Validation failed
        data:
          type: object
          additionalProperties:
            type: string
          example:
            email: Format email tidak valid!
            password: Password minimal terdiri dari 6 karakter!
    PagingResponse:
      type: object
      properties:
        currentPage:
          type: integer
          example: 0
        totalPages:
          type: integer
          example: 1
        size:
          type: integer
          example: 10
        totalElements:
          type: integer
          format: int64
          example: 5

    # Entity Response Objects
    UserResponse:
      type: object
      properties:
        idUser:
          type: integer
          format: int64
          example: 1
        nama:
          type: string
          example: Ahmad Mahasiswa
        email:
          type: string
          format: email
          example: mahasiswa@lab.com
        role:
          $ref: '#/components/schemas/Role'
        createdAt:
          type: string
          format: date-time
          example: "2026-05-18T08:50:00Z"
    WebResponseUserResponse:
      type: object
      properties:
        code:
          type: integer
          example: 201
        status:
          type: string
          example: CREATED
        message:
          type: string
          example: Registrasi pengguna berhasil dilakukan!
        data:
          $ref: '#/components/schemas/UserResponse'
    JwtResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJtYWhhc2lzd2FAbGFiLmNvbSIsImlhdCI6MTcxNTk5NjQwMCwiZXhwIjoxNzE2MDgyNDAwfQ...
        user:
          $ref: '#/components/schemas/UserResponse'
    WebResponseJwtResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        status:
          type: string
          example: OK
        message:
          type: string
          example: Login berhasil!
        data:
          $ref: '#/components/schemas/JwtResponse'
...
    BarangResponse:
      type: object
      properties:
        idBarang:
          type: integer
          format: int64
          example: 1
        idKategori:
          type: integer
          format: int64
          example: 1
        namaKategori:
          type: string
          example: Alat Ukur Elektronik
        namaBarang:
          type: string
          example: Digital Multimeter Sanwa CD800a
        kodeBarang:
          type: string
          example: BRG-EL-001
        jumlahTotal:
          type: integer
          example: 15
        jumlahTersedia:
          type: integer
          example: 13
        kondisi:
          $ref: '#/components/schemas/KondisiBarang'
        lokasi:
          type: string
          example: Lab Elektronika, Lemari B, Rak 3
    WebResponseBarangResponse:
      type: object
      properties:
        code:
          type: integer
          example: 201
        status:
          type: string
          example: CREATED
        message:
          type: string
          example: Barang inventaris berhasil didaftarkan!
        data:
          $ref: '#/components/schemas/BarangResponse'
    WebResponseListBarangResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        status:
          type: string
          example: OK
        message:
          type: string
          example: Berhasil mendapatkan daftar barang inventaris!
        data:
          type: array
          items:
            $ref: '#/components/schemas/BarangResponse'
        paging:
          $ref: '#/components/schemas/PagingResponse'
...