# src/main/java/com/inventorilab/controller/AuthController.java
package com.inventorilab.controller;

import com.inventorilab.dto.request.LoginRequest;
import com.inventorilab.dto.request.RegisterRequest;
import com.inventorilab.dto.response.JwtResponse;
import com.inventorilab.dto.response.UserResponse;
import com.inventorilab.response.WebResponse;
import com.inventorilab.service.interfaces.AuthService;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/auth")
@RequiredArgsConstructor
public class AuthController {

    private final AuthService authService;

    @PostMapping("/register")
    public ResponseEntity<WebResponse<UserResponse>> register(@Valid @RequestBody RegisterRequest request) {
        UserResponse userResponse = authService.register(request);
        WebResponse<UserResponse> response = WebResponse.<UserResponse>builder()
                .code(HttpStatus.CREATED.value())
                .status("CREATED")
                .message("Registrasi pengguna berhasil dilakukan!")
                .data(userResponse)
                .build();
        return ResponseEntity.status(HttpStatus.CREATED).body(response);
    }

    @PostMapping("/login")
    public ResponseEntity<WebResponse<JwtResponse>> login(@Valid @RequestBody LoginRequest request) {
        JwtResponse jwtResponse = authService.login(request);
        WebResponse<JwtResponse> response = WebResponse.<JwtResponse>builder()
                .code(HttpStatus.OK.value())
                .status("OK")
                .message("Login berhasil!")
                .data(jwtResponse)
                .build();
        return ResponseEntity.ok(response);
    }
}